Removed useless int

+++ WaveEditor/WaveEditor.Resize.Lib/AudioProcessor.cs
@@ -100,7 +100,7 @@ namespace WaveEditor.Resize.Lib
 
         private double Interpolate(double y0, double y1, double x0, double x1, double x)
         {
-            return (int)(y0 + (y1 - y0) * (x - x0) / (x1 - x0));
+            return (y0 + (y1 - y0) * (x - x0) / (x1 - x0));
         }
     }
 }
